requirement.txt
flask
python-dotenv
requests
faiss-cpu
langchain_community